+++ print-udp.c
@@ -354,8 +354,8 @@ udpipaddr_print(netdissect_options *ndo, const struct ip *ip, int sport, int dpo
 		} else {
 			if (sport != -1) {
 				ND_PRINT("%s > %s: ",
-					udpport_string(ndo, sport),
-					udpport_string(ndo, dport));
+					udpport_string(ndo, (uint16_t)sport),
+					udpport_string(ndo, (uint16_t)dport));
 			}
 		}
 	} else {
